feat(StationService): 노선에 등록된 역 삭제 금지 기능

+++ src/main/java/subway/service/StationService.java
@@ -2,6 +2,7 @@ package subway.service;
 
 import subway.ErrorMessage.DataErrorMessage;
 import subway.domain.Station;
+import subway.domain.StationAddedRepository;
 import subway.domain.StationRepository;
 import subway.view.SubwayOutputView;
 
@@ -15,6 +16,10 @@ public class StationService {
     }
 
     public static void deleteStation(String station) {
+        if (StationAddedRepository.isStationAddedLine(station)) {
+            throw new IllegalArgumentException(DataErrorMessage.MUST_NOT_BE_ADDED_TO_LINE.getMessage());
+        }
+
         if (!StationRepository.deleteStation(station)) {
             throw new IllegalArgumentException(DataErrorMessage.MUST_BE_EXISTING_STATION.getMessage());
         }
